remove "vote now" etc. buttons watching (illionas)
until now, a user could know whether another user has already voted or not because the "vote now" button changed to "change vote". Due to a bug it showed this behaviour not for the logged in user but for the user being watched. This commit removes the buttons entirely when watching another user.